// src/types/CheckoutInterface.ts
export interface CheckoutCompanyDetailsInterface {
  ReturnChallanNumber: string | null;
  SaleChallanNumber: string | null;
  FocChallanNumber: string | null;
  CompanyFromPanNumber: string | null;
  CompanyToPanNumber: string | null;
  CompanyFrom: string;
  CompanyTo: string;
  CompanyFromAddress: string;
  CompanyToAddress: string;
  CompanyFromPhoneNumber: string;
  CompanyToPhoneNumber: string;
  CompanyFromId: string;
  CompanyToGUId: string;
}
// ...
export interface SelectedInventoryDetailInterface {
  InventoryId: string;
  ExpiryDate: string;
  PartNumber: string;
  BatchNumber: string;
  ContainerType: string;
  PackSize: string;
  Quantity: number;
}

export interface CheckoutInventoryDescriptionInterface {
  Description: string;
  Quantity: string;
  ShortName: string;
  SelectedInventories: SelectedInventoryDetailInterface[];
}